I2util: Added conf file handling routines, hex encode/decode, md5, readpassphrase...
Basically lots of things that owamp and bwctl will both use.

Modified bwctl to use the I2util versions of these functions.

Modified bwctl to do aeskeys with either a keyfile or a passphrase.

Modified bwctld to do aeskeys with a keyfile.

(These changes have not even been compiled yet since I did this on my
laptop. I will be checking in updates to this after actually testing it.)

jeff


git-svn-id: https://svn.internet2.edu/svn/bwctl/trunk@103 2da7661c-602c-0410-9277-d9994c0c2ef1