Preparing for new release. Unfortunately, I will not be able to test until next week.
CHANGES
    enumerate changes from last release

Modified the BWLCreateContext function to be a varargs function so
Context values can be set upon creation. This is how a new variable
can be added to modify how the use of NTP is initialized.

corrected compiler warnings...

Began to document NTP changes - this is not done yet.

jeff


git-svn-id: https://svn.internet2.edu/svn/bwctl/trunk@308 2da7661c-602c-0410-9277-d9994c0c2ef1