Migrated the reset of bwctl to I2Addr's.
The code from GSoC06 is not as clean as I would like. It is going to
take some work to fix. The main reason for going to an api interface
for the tester is to do most of the testers work during test-request
negotiation. That way, things that fail can fail before they effect
the scheduling on the client side. That was not done - so there is
a fair amount of work. This was not done for the GSoC project - most
of the pieces are here - but they need to be refactored to make this
a stable/supportable piece of code.

Also, there as been a lot of bit-rot in the past 2 years since the last
release. The same type-punning issues that took me a couple of days on
owamp need to be fixed here.

Additionally, there are some problems with the way interrupt-handlers
are done.


git-svn-id: https://svn.internet2.edu/svn/bwctl/branches/BWL06CLEANUP@375 2da7661c-602c-0410-9277-d9994c0c2ef1