Removed all time functions that do not go through the BWLGetTimestamp functions. (This allows offset debugging to work...)
Corrected all type-punning except with regard to the socket API. (And
still corrected some of them - the easy ones.)
So... I'm not compliant to the ISO standard, but I am to the POSIX
standard... Hmm... Who wins?)

Added some overflow protection to 64-bit arithm functions.

Documented the BWCL_DEBUG_TIMEOFFSET environment variable. (This can
be used to fake bwctl into reading the clock as a fixed offset from
where it really is. This is very useful when debugging issues with
offset clocks.)

jeff



git-svn-id: https://svn.internet2.edu/svn/bwctl/branches/BWL06CLEANUP@388 2da7661c-602c-0410-9277-d9994c0c2ef1