Working on completely changing how fedemp did tool negotiation...
(He had it doing things in the client - which makes no sense, since the client
is not even involved in the test in 3-party.)

This required putting more things in the protocol communication.

This does not work yet, but it is getting much closer. Basically - I still
need to decide where to put the 'tool' detection portion. I'm thinking at
daemon start-up only. But, then if someone installs a new tool (like iperf)
then the daemon would need to be restarted to make it available.

(Since the tool detection is not currently running - all tests currently
fail since the tool variable is not initialized. But, I have enough
changes here that are working - I wanted to checkpoint this work.)

jeff



git-svn-id: https://svn.internet2.edu/svn/bwctl/branches/BWL06CLEANUP@402 2da7661c-602c-0410-9277-d9994c0c2ef1