Modified the Context functions a bit to remove circular dependency issues.
The context holds the error-reporting configuration, and is used
when parsing config files. However, over time more configuration
initialization (with tools and NTP specifically) happened there.

Now that options have been added into the config file to modify how
some of those things are initialized, it was necessary to split the
context creation into two parts - the first does error configuration
and can then be used for parsing the files and setting the other
configurable options.

Modified Aaron's ntp changes a bit - if the system call ever fails,
it will never be called again. It treats the system as if the timex.h
header file was not available.

Also fixed a bug in bwctl where the context was being passed in for
arg one of I2AddrByNode instead of the I2ErrHandle.

Also did a bit of prep to start moving the daemon event loop into the
library to make it easier for the client to run tools in the same
way as the server.



git-svn-id: https://svn.internet2.edu/svn/bwctl/branches/BWL06CLEANUP@420 2da7661c-602c-0410-9277-d9994c0c2ef1