Removed unneeded options from bwctld:     datadir/verbose
Removed peerport initialization process. It was overkill, and confusing.
Now, it just uses a random port in the range, and cycles through until
it can open one.

Moved many of the conf-file parsing options into a new function in
daemon.c. This will be called by both the server and the client
for implementing request-manager functionality.

Reset some of the fuzz-time constants to smaller values - now that there
is better process control on iperf, hopefully it will work to tighten
up the schedule.

Added a Memory-register function into the context. This is poor-man's
memory management. When the context is free'd, and registered pointers
are also free'd.

Removed schedule.c. It was never used in bwctl - it was old owamp
code that somehow stayed part of bwctl all this time.

Still don't have the client running tests directly, but things are now
very close.



git-svn-id: https://svn.internet2.edu/svn/bwctl/branches/BWL06CLEANUP@430 2da7661c-602c-0410-9277-d9994c0c2ef1