bwctld.c:
Changing 'fuzz' calculation in scheduling to only multiply the rtt, not the
time error.

Added debugging output for the scheduling process.

Modified the give-up 'timeout' for the end of test to use the dieby timeout 
instead of the generic controltimeout.

bwctld.limits:

Added an example of max_time_error into the file. (A daemon can say that it
will not support peers with greater than this amount of time error - including
the error on the local system.) This is intended to protect bwctld
from a possible DOS attack by clients requesting tests and specifying very
large time errors.

time.c:

Modify the way the time error is specified in the context. Used to be a
(double *), now it is just a double directly. (Required changes to the
context code to support double in the 'union' and context access functions.

endpoint.c:

Modified the poor synchronized clock detection code to make the test
fail. Before it reported the error and continued to try and work - but
it did not work.

bwctl.c:

Modified the scheduling request code to ask for a test a little closer
to current-time.

Added more schedule debugging (with -vv option specified).




git-svn-id: https://svn.internet2.edu/svn/bwctl/branches/BWL06CLEANUP@433 2da7661c-602c-0410-9277-d9994c0c2ef1