Adding a client config file (~/.bwctlrc).
This file can be used to set many tool specific aspects of bwctl. Many
of the config file options are identical to the daemon - in fact it
uses most of the same parsing code.

Also modified the 'ports' parsing code to allow the specification of
a single port. (and modified the tool parsing options to use the keyword
'port' instead of 'ports' to make it compatible with previous versions
of bwctl - this was just a mistake on my part when I abstracted the
tools)

Modified the bwctl.man page to indicate the BWCTLRC env variable, and
removed the now defunct environment variables that used to be used to
set these daemon-features.

(Still need to modifiy the client to use the same functions as the daemon
to run the tool - but now that the configuration is available in the
'ctx' variable, this will be easy.)

jeff


git-svn-id: https://svn.internet2.edu/svn/bwctl/branches/BWL06CLEANUP@435 2da7661c-602c-0410-9277-d9994c0c2ef1